Problema inesperado: A centralização da rotação estava fazendo que ao girar uma imagem, as demais fossem movimentadas também. Também perdi tempo por nomear porcamente as variaveis. (É melhor trocar o sufixo de variaveis que o prefixo)

+++ src/Vector2.h
@@ -46,13 +46,13 @@ public:
 
    Vector2 operator - (const Vector2& v)
    {
-        Vector2 aux( x - v.x, y - v.y);
+        Vector2 aux( x - (numeric)v.x, y - (numeric)v.y);
         return( aux );
    }
 
    Vector2 operator + (const Vector2& v)
    {
-       Vector2 aux( x + v.x, y + v.y);
+       Vector2 aux( x + (numeric)v.x, y + (numeric)v.y);
        return( aux );
    }
